Clean up 21 gene panel
I increased the score from 37 to 47 to avoid any ambiguity with
interpreting it as "high". The 37 is closer to the NCCN guideline cutoff,
and the interpretation of these scores takes into account a number of
other factors (e.g., patient's age). Getting the score in the example
away from the cutoff avoids any ambiguity with interpreting as "high".