Remove (commented out) inline extension
The extension for EvidenceType is only used in one place, so Mark tried changing it to an inline extension so we could have one less StructureDefinition. However, the IG Publisher raised an error that the URL for this extension had to be an absolute URL. The URL of inline extensions are like just names, like "evidenceType". So Mark reverted back to a Standalone extension.

This commit cleans up the leftover inline extension code.