Abstract slicing rules into rule sets
We were not setting Category 0..* correctly. Now following the correct pattern that allows additional categories.
Chat log with Chris:
So this goes for any field that is multi-value (e.g., card > 1)...  If you just do straight assignment, you are saying that every value needs to match.  So you need to slice it on $this by value or pattern (depending on if you use fixed[x] or pattern[x] in the assignment) -- and then define a slice w/ the assignment you want.

Here is an example of exactly this in US Core: http://www.hl7.org/fhir/us/core/StructureDefinition-us-core-observation-lab.html (edited)