Clean up search param types, move _id to 3rd fallback
This commit fixes a bug caused by a mistake with the indices for
`searchParam` elements, where multiple `searchParams` were assigned
to `[0]` for the same resource, causing silent overwriting.

As part of this, I realized that I could move the `_id` for `Patient`
just to the 3rd fallback. Previously `GET [base]/Patient?_id=...`
support was required for all CapabilityStatements. However, it's only
strictly necessary for the 3rd fallback.