Create new parent for TNM Categories
The max value set approach is triggering what I believe to be an IG Publisher bug, where the profile conformance is not looking at the maximum value sets and simply assumes because a value set is preferred, any value will do. This is causing new warning messages that TNM categories are matching multiple slices (which they do not -- the max value sets do not overlap at all). Should we work around this or live with the spurious warnings?